Refactor: Move DetailsCard, update dependencies, and improve UI interactions
- Moved `DetailsCard.kt` from `paginainicial.ui.components` to `historico.ui.components`.
- Moved `DadosContaUI.kt` to `adicionarconta.models` package.
- Updated `gradle/libs.versions.toml` with newer library versions.
- Updated `gradle/wrapper/gradle-wrapper.properties` to use Gradle 8.14.2.
- In `HistoricoViewModel.kt`:
    - Removed unused `_contasEncontradas` map.
    - Added `ToastManager` to display a toast when no accounts are found for the current month.
- In `ParcelamentoColumn.kt`:
    - Improved formatting of the "Parcelado em" text to correctly display information when interest is applied.
- In `BreezeBottomBar.kt`:
    - Refactored to derive `selectedItem` from `navController.currentBackStackEntryAsState()` to ensure correct item selection upon navigation.
    - Removed `BreezeBottomBarModel` and its preview.
- Removed unused imports and variables across multiple files.
- Commented out `calculaValorTotalConta` in `Utils.kt`.
- Introduced `ToastManager` in `Utils.kt` to manage toast display intervals.
- In `app/build.gradle.kts`, updated `versionName` to "1.2-beta".
