+++ app/src/main/java/com/migueldk17/breeze/ui/features/adicionarconta/ui/layouts/Final.kt
@@ -29,7 +29,7 @@ import com.migueldk17.breeze.ui.components.DescriptionText
 import com.migueldk17.breeze.ui.features.adicionarconta.ui.components.PersonalizationCard
 import com.migueldk17.breeze.ui.features.adicionarconta.viewmodels.AdicionarContaViewModel
 import com.migueldk17.breeze.ui.features.historico.ui.components.retornaValorTotalArredondado
-import com.migueldk17.breeze.ui.features.paginainicial.ui.components.DetailsCard
+import com.migueldk17.breeze.ui.features.historico.ui.components.DetailsCard
 import com.migueldk17.breeze.ui.features.paginainicial.ui.components.avançaMainActivity
 import com.migueldk17.breeze.ui.utils.formataSaldo
 import com.migueldk17.breeze.ui.utils.formataTaxaDeJuros
