+++ app/src/main/java/com/migueldk17/breeze/ui/features/adicionarconta/viewmodels/AdicionarContaViewModel.kt
@@ -12,7 +12,7 @@ import com.migueldk17.breeze.entity.Conta
 import com.migueldk17.breeze.entity.ParcelaEntity
 import com.migueldk17.breeze.repository.ContaRepository
 import com.migueldk17.breeze.repository.ParcelaRepository
-import com.migueldk17.breeze.ui.features.adicionarconta.DadosContaUI
+import com.migueldk17.breeze.ui.features.adicionarconta.models.DadosContaUI
 import com.migueldk17.breeze.ui.utils.arredondarValor
 import com.migueldk17.breeze.uistate.UiState
 import dagger.hilt.android.lifecycle.HiltViewModel
