+++ app/src/main/java/com/migueldk17/breeze/ui/utils/Utils.kt
@@ -1,6 +1,8 @@
 package com.migueldk17.breeze.ui.utils
 
 
+import android.content.Context
+import android.widget.Toast
 import androidx.navigation.NavController
 import java.math.BigDecimal
 import java.math.RoundingMode
@@ -49,22 +51,22 @@ fun formataMesAno(localDate: LocalDate): String {
     return "%04d-%02d".format(localDate.year, localDate.monthValue)
 }
 
-fun calculaValorTotalConta(
-    valorParcela: Double,
-    taxaJurosMensal: Double,
-    qtdParcelas: Int
-): Double {
-    if (qtdParcelas <= 0 || taxaJurosMensal < 0) return 0.0
-
-    val i = taxaJurosMensal
-    val n = qtdParcelas
-
-    return if (i == 0.0){
-        valorParcela * n
-    } else {
-        valorParcela * (1 - (1 + i).pow(-n)) / i
-    }
-}
+//fun calculaValorTotalConta(
+//    valorParcela: Double,
+//    taxaJurosMensal: Double,
+//    qtdParcelas: Int
+//): Double {
+//    if (qtdParcelas <= 0 || taxaJurosMensal < 0) return 0.0
+//
+//    val i = taxaJurosMensal
+//    val n = qtdParcelas
+//
+//    return if (i == 0.0){
+//        valorParcela * n
+//    } else {
+//        valorParcela * (1 - (1 + i).pow(-n)) / i
+//    }
+//}
 fun formataTaxaDeJuros(taxaDeJurosMensal: Double): String {
     val formato = DecimalFormat("#.##%") // Define a máscara: duas casas decimais e símbolo de porcentagem
     return formato.format(taxaDeJurosMensal)
@@ -79,6 +81,19 @@ fun NavController.navigateSingleTopTo(route: String) {
         restoreState = true
     }
     }
+
+object ToastManager {
+    private var lastToastTime = 0L
+    private val toastInterval = 2000
+
+    fun showToast(context: Context, message: String) {
+        val currentTime = System.currentTimeMillis()
+        if (currentTime - lastToastTime > toastInterval) {
+            Toast.makeText(context, message, Toast.LENGTH_SHORT).show()
+            lastToastTime = currentTime
+        }
+    }
+}
