+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.6.1"
+agp = "8.11.0"
 breezeIcons = "v1.0.10"
 kotlin = "2.0.21"
 coreKtx = "1.16.0"
@@ -10,16 +10,16 @@ lifecycleRuntimeKtx = "2.9.1"
 activityCompose = "1.10.1"
 composeBom = "2025.06.01"
 material3 = "1.4.0-alpha16"
-googleIcons = "1.6.0"
+googleIcons = "1.7.8"
 navComposeVersion = "2.9.0"
 roomVersion = "2.7.2"
 ksp = "2.0.21-1.0.25"
-dagger-hilt = "2.51.1"
+dagger-hilt = "2.56.2"
 hiltNavigationCompose = "1.2.0"
 composeUi = "1.8.3"
-airbnbLottie = "6.6.2"
-googleServices = "4.4.2"
-firebaseBom = "33.15.0"
+airbnbLottie = "6.6.7"
+googleServices = "4.4.3"
+firebaseBom = "33.16.0"
 pluginCrashlyticsVersion = "3.0.4"
 
 [libraries]
