+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,5 +1,5 @@
 #Mon Nov 11 18:30:42 BRT 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-8.7-bin.zip
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.14.2-bin.zip
